fix: altitude of type uint8 instead of int

+++ src/controller/controller/lib/helper/helper.h
@@ -15,7 +15,7 @@ struct Collection {
   int pitch;
   int yaw;
   int throttle;
-  int altitude;
+  uint8_t altitude;
   byte FLAGS;   // bit‑mapped flags
 };
